fix: updates the template to use the new dashboard definition in sitemap
